Rename tests/utils/{mock_microgrid.py => mock_microgrid_client.py}
