+++ tests/utils/component_data_streamer.py
@@ -11,7 +11,7 @@ from datetime import datetime, timezone
 from frequenz.sdk._internal.asyncio import cancel_and_await
 from frequenz.sdk.microgrid.component import ComponentData
 
-from .mock_microgrid import MockMicrogridClient
+from .mock_microgrid_client import MockMicrogridClient
 
 
 class MockComponentDataStreamer:
