+++ tests/utils/mock_microgrid_client.py
